% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBesselI.R, R/cBesselI_next.R
\name{cBesselI}
\alias{cBesselI}
\alias{cBesselI_next}
\title{Continued BesselI function}
\usage{
cBesselI(zs, v)

cBesselI_next(z1, z0, branch0, v, expon_scaled = FALSE)
}
\arguments{
\item{zs}{vector of complex or real numbers.}

\item{v}{numeric (scalar), i.e., the order (maybe fractional and
negative) of the Modified Bessel function of the first kind, \eqn{v} in
\eqn{I_v(z)}.}

\item{z1}{current input complex number.}

\item{z0}{previous input complex number.}

\item{branch0}{branch for z0.}

\item{expon_scaled}{whether the BesselI() should be scaled.}
}
\value{
\itemize{
\item \code{cBesselI()}: vector of complex or real numbers.
\item \code{cBesselI_next()}: list of current input's function value and branch.
}
}
\description{
Accomplish the continuity of the BesselI (the modified Bessel function
of the first kind) function. see \code{\link[=CF]{CF()}}.
The following continuation formula from Abramowitz and Stegun (1972):
\deqn{
  I_v(ze^{m\pi i}) = e^{mv\pi i} I_v(z), \{m~~\text{integer}\},
}
is used to calculate the value of \eqn{I_v(z)} for \eqn{arg(z)} different
than its principle value.
}
\examples{
# a = seq(0, 1500, 0.001)
a = seq(0, 1500, 100)
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
ga = sqrt(k^2 - 2*sigma^2*a*1i)
dk = exp(-k*tau); da = exp(-ga*tau)
zs = sqrt(v0*v1) * (4*ga/sigma^2) * exp(-0.5*ga*tau)/(1-da)
nu = 2*theta*k/sigma^2 - 1
f = cBesselI(zs, nu)
par(mfrow=c(2,1))
plot(Re(zs), Im(zs), col="blue", type="l"); abline(h=0, v=0, col="red")
plot(Re(f), Im(f), col="blue", type="l", main="Continued BesselI function")

z1 = -1 - 0.1i; z0 = -1 + 0.1i; branch0 = 0
f_branch = cBesselI_next(z1, z0, branch0, nu)
}
