% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF.R, R/CF_next.R
\name{CF}
\alias{CF}
\alias{CF_next}
\title{Characteristic Function of IV}
\usage{
CF(a, v0, v1, tau, k, theta, sigma)

CF_next(a1, z0, branch0, v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{a}{vector of real numbers, does't work for a single number,
should also work for complex numbers.}

\item{v0}{value of the left end, \eqn{v_u}.}

\item{v1}{value of the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{a1}{current input number, scalar, complex number should work still.}

\item{z0}{previous input for the embedded function \code{cBesselI_next()}.}

\item{branch0}{branch of z0.}
}
\value{
\itemize{
\item \code{CF()}: vector of complex numbers.
\item \code{CF_next()}: list of (CF_next(a1), z1, branch1).
}
}
\description{
The conditional Characteristic Function (CF) of the integrated variance,
given the two side levels of the variance.
\itemize{
\item \code{CF()}: Compute the CF values all at once.
\item \code{CF_next()}: Compute the CF value one by one.
}
}
\details{
The integrated variance over period \eqn{(u,t]} is defined as
\deqn{
  IV_{u,t} \triangleq  \int_{u}^{t}v(s)ds.
}
The conditional CF is defined as
\eqn{
  \Phi(a) \triangleq  E\left[ e^{iaIV_{u,t}}|v_u,v_t \right],
}
which can be decomposed into three parts:
\eqn{
  \Phi(a) = f_1(a) \times f_2(a) \times f_3(a),
}
where
\deqn{
  f_1(a) \triangleq
  \frac{\gamma(a)}{k} e^{-0.5(\gamma(a)-k)(t-u)}
  \frac{1-e^{-k(t-u)}}
     {1-e^{-\gamma(a)(t-u)}},
}
\deqn{
  f_2(a) \triangleq
   \text{exp}\left\{ \frac{v_u+v_t}{\sigma^2}\left[
    k\frac{1+e^{-k(t-u)}}{1-e^{-k(t-u)}} -
    \gamma(a)\frac{1+e^{-\gamma(a)(t-u)}}{1-e^{-\gamma(a)(t-u)}}
  \right] \right\},
}
\deqn{
  f_3(a) \triangleq
   I_{0.5d-1}\left[
     \sqrt{v_uv_t}\frac{4\gamma(a)}{\sigma^2} \frac{e^{-0.5\gamma(a)(t-u)}}
         {1-e^{-\gamma(a)(t-u)}} \right]
                     \Big/I_{0.5d-1}\left[
     \sqrt{v_uv_t}\frac{4k}{\sigma^2} \frac{e^{-0.5k(t-u)}}
         {1-e^{-k(t-u)}} \right],
}
\eqn{\gamma(a) = \sqrt{k^2 - 2\sigma^2ai}}, \eqn{d = 4\theta k/\sigma^2},
and \eqn{I_v(\cdot)} is the modified Bessel function of the first kind.
\itemize{
\item The evaluation of \eqn{I_v(z)} with a complex number input is done with
\code{\link[Bessel:Bessel]{Bessel::BesselI()}} or \code{\link[Bessel:besselIasym]{Bessel::besselIasym()}}, the numerator in \eqn{f_3(a)}.
\item The evaluation of \eqn{I_v(x)} with a real number input is done with
\code{\link[Bessel:Bessel]{Bessel::BesselI()}} or \code{\link[Bessel:besselIasym]{Bessel::besselIasym()}}, the denominator in \eqn{f_3(a)}.
}

The modified Bessel function of the first kind is given as the following
power series:
\deqn{
  I_v(z) = \left(\frac{1}{2}z\right)^v \sum_{j=0}^{\infty}
  \frac{\left(\frac{1}{4}z^2\right)^j}{j!\Gamma(v+j+1)},
}
where \eqn{\Gamma(x)} is the gamma function and \eqn{z} is a complex number.
}
\examples{
# a = seq(0, 1500, 0.001)
a = seq(0, 1500, 100)
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
f = CF(a, v0, v1, tau, k, theta, sigma)
par(mfrow = c(2,1))
plot(Re(f), Im(f), type="l", col="blue")
plot(a, Re(f), type="l", main="Real part of the CF"); abline(h=0, col="red")
}
\references{
Broadie, M., & Kaya, Ö. (2006). Exact simulation of stochastic volatility
and other affine jump diffusion processes. Operations research,
54(2), 217-231.
}
