% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_trunc.R
\name{CF_trunc}
\alias{CF_trunc}
\title{Truncate the Characteristic Function}
\usage{
CF_trunc(h, v0, v1, tau, k, theta, sigma, eps = 1e-05)
}
\arguments{
\item{h}{grid size.}

\item{v0}{the left side variance level, \eqn{v_u}.}

\item{v1}{the right side variance level, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{eps}{truncate error bound.}
}
\value{
vector of CF points.
}
\description{
Truncate the Characteristic Function automatically through setting up the
truncation error bound.
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
mu = miv(v0, v1, tau, k, theta, sigma); sd = sqrt(mu[2] - mu[1]^2)
mu_eps = mu[1] + 7*sd; h = pi/mu_eps
zs = CF_trunc(h, v0, v1, tau, k, theta, sigma)
plot(Re(zs), Im(zs), type="b", col="blue"); abline(h=0,v=0, col="red")
hj = seq(1, length(zs)) * h
plot(hj, Re(zs), type="b", col="blue"); abline(h=0, col="red")
}
