% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiv.R
\name{ddiv}
\alias{ddiv}
\title{Derivative of the Density of IV}
\usage{
ddiv(x, v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{x}{vector of quantile.}

\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
derivative/derivatives corresponding to \eqn{x}, depending on
whether the supplied \eqn{x} is a scalar or a vector.
}
\description{
Derivative of the conditional Density of the Integrated Variance (IV).
Use it carefully, because currently the truncation of Fourier inversion of
the CF is not done in a robust way.
}
\examples{
# x = seq(0,0.1,0.001)
# k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1

# dd = ddiv(x, v0, v1, tau, k, theta, sigma)
# txt = "Derivative of Conditional PDF of IV"
# plot(x, dd, type="l", col="blue", ylab="Derivative", main=txt)
# abline(h=0,v=0)
}
