% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branch.R, R/get_branch_next.R
\name{get_branch}
\alias{get_branch}
\alias{get_branch_next}
\title{Determine the correct branch}
\usage{
get_branch(zs)

get_branch_next(z1, z0, branch0)
}
\arguments{
\item{zs}{vector of complex numbers.}

\item{z1}{current complex number.}

\item{z0}{previous complex number.}

\item{branch0}{branch of z0.}
}
\value{
\itemize{
\item \code{get_branch()}: vector of floors of the branches.
\item \code{get_branch_next()}: scalar of floor of the current branch.
}
}
\description{
To accomplish the continuity of the Characteristic Function, we need to
track of arg(z) and change the branch when necessary, i.e., when the
complex numbers cross the negative real axis. see \code{\link[=CF]{CF()}} and
\code{\link[=cBesselI]{cBesselI()}}.
\itemize{
\item \code{get_branch()} gets the branches all at once.
\item \code{get_branch_next()} gets the branch one by one.
}
}
\examples{
# a = seq(0, 1500, 0.001)
a = seq(0, 1500, 100)
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
ga = sqrt(k^2 - 2*sigma^2*a*1i)
dk = exp(-k*tau); da = exp(-ga*tau)
zs = sqrt(v0*v1) * (4*ga/sigma^2) * exp(-0.5*ga*tau)/(1-da)
par(mfrow = c(1,1))
plot(Re(zs), Im(zs), col="blue")
branches = get_branch(zs)
#
z1 = -1 - 0.1i; z0 = -1 + 0.1i; branch0 = 0
branch1 = get_branch_next(z1, z0, branch0)
}
