% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miv.R
\name{miv}
\alias{miv}
\title{First Two Conditional Moments of IV}
\usage{
miv(v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{v0}{the left side variance level, \eqn{v_u}.}

\item{v1}{the right side variance level, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
vector of the first two un-centralized moments \eqn{\mu_1, \mu_2}.
}
\description{
Compute the first two conditional moments of IV, given \eqn{v_u} and
\eqn{v_t}.
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
mu = miv(v0, v1, tau, k, theta, sigma)
sd = sqrt(mu[2] - mu[1]^2)
x = seq(mu[1]-4*sd, mu[1]+4*sd, 0.001)
plot(x, dnorm(x, mu[1], sd), type="l", col="blue")
}
