% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R, R/piv.R, R/riv.R
\name{div}
\alias{div}
\alias{piv}
\alias{riv}
\title{Conditional Distribution of IV}
\usage{
div(x, v0, v1, tau, k, theta, sigma)

piv(q, v0, v1, tau, k, theta, sigma)

riv(n, v0, v1, tau, k, theta, sigma, log.inv = FALSE)
}
\arguments{
\item{x}{vector of quantile.}

\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{q}{vector of quantile.}

\item{n}{number of observations.}

\item{log.inv}{\code{FALSE}, logging the Fourier inversion if \code{log.inv = TRUE}.}
}
\value{
\code{div} gives the density, \code{piv} gives the distribution function, and
\code{riv} generates random deviates.
}
\description{
Density, distribution function and random variable generation for the
conditional distribution of the Integrated Variance (IV),
all through Fourier inversion of the Characteristic Function,
see \code{\link[=CF]{CF()}} for the details.

Currently, only \code{\link[=riv]{riv()}} has been tested heavily and can be used
safely for most scenarios.

While, for other functions,
use them carefully, because currently the truncation of Fourier inversion of
the CF is not done in a robust way.
}
\details{
The distribution function is inversed from the conditional CF, see
\code{\link[=invCF]{invCF()}} and \code{\link[=CF]{CF()}}.
Noting that \code{piv()} is a wrapper for \code{invCF()}.
Currently, the discretized inversions are done through:
\itemize{
\item \code{piv()}, \code{div()}:  \code{hj = seq(0.001, 1500, 0.001)}, i.e., \code{h=0.001}
and truncated at \code{j=length(hj)}.
\item \code{riv()}: \eqn{h = \pi/\mu_{\epsilon}} with
\eqn{\mu_{\epsilon} = \mu_{iv} + 7\sigma_{iv}}, where \eqn{\mu_{iv}} and
\eqn{\sigma_{iv}} denote the conditional mean and standard deviation of the
integrated variance, respectively. Meanwhile, the truncation is done at
\eqn{j=N} when
\deqn{
  \frac{|\Phi(hN)|}{N} < \frac{\pi\epsilon}{2},
}
where \eqn{\epsilon} is the desired truncation error.
}
}
\examples{
# x = seq(0,0.1,0.001)
# k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1

# Pr = piv(x, v0, v1, tau, k, theta, sigma)
# par(mfrow=c(1,1))
# txt = "Conditional CDF of IV"
# plot(x,Pr,type="l",col="blue",ylab="CDF",main=txt)
# abline(h=0,v=0); abline(h=1, col="red")

# d = div(x, v0, v1, tau, k, theta, sigma)
# txt = "Conditional PDF of IV"
# plot(x, d, type="l", col="blue", ylab="PDF", main=txt); abline(h=0,v=0)

# r = riv(1, v0, v1, tau, k, theta, sigma)

}
\references{
Broadie, M., & Kaya, Ö. (2006). Exact simulation of stochastic volatility
and other affine jump diffusion processes. Operations research,
54(2), 217-231.
}
\seealso{
See \code{\link[=CF]{CF()}} for the details about the Characteristic
Function.
}
