% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_hest.R
\name{price_hest}
\alias{price_hest}
\title{Pricing the European Call Option Under the Heston Model}
\usage{
price_hest(N, S, K, v0, tau, r, k, theta, sigma, rho, true_price, log.inv = F)
}
\arguments{
\item{N}{number of underling asset price samples to simulate.}

\item{S}{current underling asset price.}

\item{K}{striking price.}

\item{v0}{current variance level.}

\item{tau}{maturity time.}

\item{r}{riskless rate.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{\rho}.}

\item{true_price}{theoretical price of the option, calculable from
the Heston (1993) method.}

\item{log.inv}{\code{FALSE}, logging the Fourier inversion if \code{log.inv = TRUE}.}
}
\value{
vector of pricing error and consumed time (simulation).
}
\description{
Pricing the European call option under the Heston model using
Broadie-Kaya (2006) method for the simulation.
}
\examples{
S = 100; K = 100; v0 = 0.010201; k = 6.21; theta = 0.019; sigma = 0.61
rho = -0.7; r = 0.0319; tau = 1; true_price = 6.8061
# price_hest(10000, S, K, v0, tau, r, k, theta, sigma, rho, true_price)
}
