% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\name{rv}
\alias{rv}
\title{Simulate Next Variance}
\usage{
rv(v0, tau, k, theta, sigma)
}
\arguments{
\item{v0}{current variance level.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
next variance, scalar.
}
\description{
Simulate the next variance given the current variance level.
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; tau = 1
v1 = rv(v0, tau, k, theta, sigma)
}
