% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invCF.R
\name{invCF}
\alias{invCF}
\title{Cumulative Distribution Function of IV}
\usage{
invCF(x, v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{x}{a scalar or vector of real numbers, support of the CDF.}

\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
probability/probabilities corresponding to \eqn{x}, depending on
whether the supplied \eqn{x} is a scalar or a vector.
}
\description{
Get the conditional Cumulative Distribution Function (CDF) of the
integrated variance, given the two side variance levels.
Use it carefully, because currently the truncation of Fourier inversion of
the CF is not done in a robust way.
}
\details{
The CDF is computed using Fourier inversion methods:
\deqn{
  F(x)\equiv \text{P}\{IV_{u,t} \le x\} = \frac{2}{\pi}\int_{0}^{\infty}
  \frac{\sin(ux)}{u}\text{Re}[\Phi(u)]du.
}
The trapezoidal rule is used to compute the probability distribution
numerically:
\deqn{
  \text{P}\{IV_{u,t} \le x\} \approx \frac{hx}{\pi} + \frac{2}{\pi}
  \sum_{j=1}^{\infty}\frac{\sin(hjx)}{j}\text{Re}[\Phi(hj)],
}
where \eqn{h} is the grid size, which is set as 0.001 internally, and
\code{hj = seq(0.001, 1500, 0.001)}.
}
\examples{
# x = seq(0,0.1,0.001)
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
# Pr = invCF(x, v0, v1, tau, k, theta, sigma)
# par(mfrow=c(1,1))
# txt = "Cumulative Probability Function"
# plot(x, Pr, type="l", col="blue", ylab="CDF", main=txt)
# abline(h=0, v=0); abline(h=1, col="red")
}
\references{
Broadie, M., & Kaya, Ö. (2006). Exact simulation of stochastic volatility
and other affine jump diffusion processes. Operations research,
54(2), 217-231.
}
\seealso{
See \code{\link[=CF]{CF()}} for the details about the Characteristic
Function.
}
