% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryield_Hest.R
\name{ryield_Hest}
\alias{ryield_Hest}
\title{Simulate Yield using Broadie-Kaya Method for the Heston Model}
\usage{
ryield_Hest(n, v0, tau, r, k, theta, sigma, rho, log.inv = FALSE)
}
\arguments{
\item{n}{number of yield samples to simulate.}

\item{v0}{current variance level.}

\item{tau}{time difference, \eqn{t-u}.}

\item{r}{riskless rate, risk-neutral drift.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{\rho}.}

\item{log.inv}{\code{FALSE}, logging the Fourier inversion if \code{log.inv = TRUE}.}
}
\value{
next period yield, a scalar.
}
\description{
Simulate the next period yield given current variance using the Broadie-Kaya
method.
}
\examples{
v0 = 0.010201; k = 6.21; theta = 0.019; sigma = 0.61; rho = -0.7
r = 0.0319; tau = 1
# Y = ryield_Hest(1000, v0, tau, r, k, theta, sigma, rho)
# hist(Y)
}
